+++ common/src/python/outputs/errors.py
@@ -21,6 +21,56 @@ log = logging.getLogger(__name__)
 preprocess_errors = {
     SysErrorCodes.ADCID_MISMATCH:
     "ADCID must match the ADCID of the center uploading the data",
+    SysErrorCodes.IVP_EXISTS:
+    "Only one Initial Visit Packet is allowed per participant",
+    SysErrorCodes.UDS_NOT_MATCH:
+    "Cannot find a matching UDS visit packet with the " +
+    "same visit number and visit date as the module packet",
+    SysErrorCodes.INVALID_MODULE_PACKET:
+    "Follow-up module packet cannot be submitted for a UDS initial visit packet (I)",
+    SysErrorCodes.UDS_NOT_EXIST:
+    "A UDS packet must be submitted before submitting this module/form",
+    SysErrorCodes.DIFF_VISITDATE:
+    "Two packets cannot have the same visit number (VISITNUM) " +
+    "if they are from different dates (VISITDATE)",
+    SysErrorCodes.DIFF_VISITNUM:
+    "Two packets cannot have the same visit date (VISITDATE) " +
+    "if they are from different visit numbers (VISITNUM)",
+    SysErrorCodes.LOWER_FVP_VISITNUM:
+    "Visit number (VISITNUM) for Follow-Up Visit Packets must be " +
+    "greater than Visit number (VISITNUM) for Initial Visit Packet",
+    SysErrorCodes.LOWER_I4_VISITNUM:
+    "Visit number (VISITNUM) for UDSv4 Initial Visit Packet (PACKET=I4) " +
+    "must be greater than Visit number (VISITNUM) for of the last UDSv3 Visit Packet",
+    SysErrorCodes.LOWER_FVP_VISITDATE:
+    "Follow-Up Packet visit date (VISITDATE) cannot be equal to " +
+    "or from a date before the Initial Visit Packet",
+    SysErrorCodes.LOWER_I4_VISITDATE:
+    "Visit date (VISITDATE) for UDSv4 Initial Visit Packet (PACKET=I4) " +
+    "must be a date after the last UDSv3 Visit Packet",
+    SysErrorCodes.EXCLUDED_FIELDS:
+    "Some fields in the input record do not match with the submitted version",
+    SysErrorCodes.INVALID_PACKET:
+    "Provided PACKET code is not in the list of accepted packets for this module",
+    SysErrorCodes.INVALID_VERSION:
+    "Provided FORMVER is not in the list of " +
+    "accepted versions for this module",
+    SysErrorCodes.INVALID_PTID:
+    "PTID must be an alpha numeric code " + "that is no more than 10 digits",
+    SysErrorCodes.INVALID_MODULE:
+    "Provided MODULE is not in the list of currently accepted modules",
+    SysErrorCodes.MISSING_IVP:
+    "Follow-Up visit cannot be submitted without an existing Initial Visit Packet",
+    SysErrorCodes.MULTIPLE_IVP:
+    "More than one IVP packet found for the participant/module",
+    SysErrorCodes.UDS_NOT_APPROVED:
+    "UDS visit packet must be approved before the module visit can be processed",
+    SysErrorCodes.MISSING_UDS_V3:
+    "To submit an Initial UDSv4 Visit Packet (PACKET=I4), " +
+    "participant must have an existing UDSv3 Visit Packet",
+    SysErrorCodes.MISSING_UDS_I4:
+    "Participant must have an existing Initial UDSv4 Visit Packet (PACKET=I4) "
+    + "submitted before the Follow-Up Visit Packet (PACKET=F)",
 }
 
 
@@ -204,17 +254,21 @@ def previous_visit_failed_error(prev_visit: str) -> FileError:
 
 def preprocessing_error(field: str,
                         value: str,
-                        line: int,
+                        line: Optional[int] = None,
                         error_code: Optional[str] = None,
-                        message: Optional[str] = None) -> FileError:
+                        message: Optional[str] = None,
+                        ptid: Optional[str] = None,
+                        visitnum: Optional[str] = None) -> FileError:
     """Creates a FileError for pre-processing error.
 
     Args:
       field: the field name
-      value: the unexpected value
-      line: the line number
+      value: the value
+      line (optional): the line number
       error_code (optional): pre-processing error code
       message (optional): the error message
+      ptid (optional): PTID if known
+      visitnum (optional): visitnum if known
 
     Returns:
       the constructed FileError
@@ -230,8 +284,11 @@ def preprocessing_error(field: str,
         error_type='error',
         error_code=error_code if error_code else 'preprocess-error',
         value=value,
-        location=CSVLocation(line=line, column_name=field),
-        message=error_message)
+        location=CSVLocation(line=line, column_name=field)
+        if line else JSONLocation(key_path=field),
+        message=error_message,
+        ptid=ptid,
+        visitnum=visitnum)
 
 
 def partially_failed_file_error() -> FileError: