+++ gear/form_qc_checker/src/python/form_qc_app/error_info.py
@@ -209,10 +209,7 @@ def is_composite_rule(rule: str) -> bool:
     Returns:
         bool: True if rule is in composite rules list
     """
-    if rule in COMPOSITE_RULES:
-        return True
-
-    return False
+    return rule in COMPOSITE_RULES
 
 
 class ErrorComposer():