+++ docs/identifier_lookup/CHANGELOG.md
@@ -2,12 +2,19 @@
 
 All notable changes to this gear are documented in this file.
 
-## 1.0.7 (Unreleased)
+## 1.0.9 (Unreleased)
 
 * Refactors to use `InputFileWrapper.get_parent_project`
 
-## 1.0.6
+## 1.0.8
+* Fixes CSV line number in error reports (exclude header row)
+* Updates pre-processing error codes
+* Updates GUID format and max length
+  
+## 1.0.7
+* Adds pre-processing check for PTID length
 
+## 1.0.6
 * Fixes bug where an empty file would be generated if no entries passed
 
 ## 1.0.5
@@ -15,7 +22,6 @@ All notable changes to this gear are documented in this file.
 * Updates identifier lookup error messages and identifier file suffix.
 
 ## 1.0.0
-
 * Adds ability to do a reverse lookup on NACCID to find the center IDs.
   Allows injecting the ADCID needed to split a CSV for distribution across centers.